feat: implement interface in module
