+++ utils/store-accessor.ts
@@ -1,12 +1,12 @@
 /* eslint-disable import/no-mutable-exports */
 import { Store } from 'vuex'
 import { getModule } from 'vuex-module-decorators'
-import MovieModule from '~/store/moviediscover/domain/MovieModule'
+import MovieModuleImpl from '~/store/moviediscover/domain/MovieModuleImpl'
 
-let movieModule: MovieModule
+let movieModule: MovieModuleImpl
 
 function initialiseStores(store: Store<any>): void {
-  movieModule = getModule(MovieModule, store)
+  movieModule = getModule(MovieModuleImpl, store)
 }
 
 export { initialiseStores, movieModule }
